refactor(utils): move mse_loss_sum to plot Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
